(fix) change test_bitmar_exchange for test_bitmart_exchange
